requirements.txt
streamlit>=1.28.0
textblob>=0.17.0
PyPDF2>=3.0.0
pandas>=1.5.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
plotly>=5.17.0
seaborn>=0.12.0
requests>=2.31.0
openpyxl>=3.1.0
python-docx>=0.8.11
Pillow>=10.0.0
wordcloud>=1.9.2
spacy>=3.7.0